eureka part

+++ booking-service/src/main/java/com/booking/demo/modal/SalonReport.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class SalonReport {
     private Long salonId;
     private String salonName;
-    private int totalEarnings;
+    private Long totalEarnings;
     private Integer totalBookings;
     private Integer cancelledBookings;
     private Double totalRefund;
